% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Bootstrap_Test_Unimodality_WTA}
\alias{Bootstrap_Test_Unimodality_WTA}
\title{Test for Unimodality for Single Stimuli Trials (Whole Trial Analysis)}
\usage{
Bootstrap_Test_Unimodality_WTA(
  obs_dat,
  eval_grid = NULL,
  h_grid = NULL,
  n_boot = 10000L
)
}
\arguments{
\item{obs_dat}{Vector containing number of spikes for each trial}

\item{eval_grid}{Vector containing points over which to evaluate the density (default is 500 points)}

\item{h_grid}{Vector containing a list of bandwidths for the Gaussian KDE (default is adaptively chosen)}

\item{n_boot}{Integer indicating the number of bootstrap samples to use (default is 10000)}
}
\value{
p_val Estimated p-value of test under the null hypothesis that the distribution is unimodal
}
\description{
This function conducts a bootstrap-based test for unimodality. This test is
used to confirm that the single stimulus trials are unimodal in the distribution
of spike counts.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{eval_grid}}{points should cover the range of observed spike counts plus slightly more}
  \item{\code{h_grid}}{all bandwidths should be positive}
  \item{\code{n_boot}}{must be greater than 1}
}
}

\examples{
## Load sample data 
## Note there is no difference between time homogeneous and time inhomogeneous processes
dat <- readRDS(system.file("test-data", "time_homogeneous_sample_dat.RDS", package = "NeuralComp"))

## Run test for A process
p_val_A <- Bootsrap_Test_Unimodality_WTA(dat$n_A)

## Run test for B process
p_val_B <- Bootsrap_Test_Unimodality_WTA(dat$n_B)

}
