% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Competition_Posterior_Predictive}
\alias{Competition_Posterior_Predictive}
\title{Posterior Predictive Sampling for Competition Model}
\usage{
Competition_Posterior_Predictive(
  end_time,
  basis_degree,
  boundary_knots,
  internal_knots,
  Results,
  burnin_prop = 0.2,
  time_inhomogeneous = TRUE,
  n_samples = 10000L
)
}
\arguments{
\item{basis_degree}{Integer indicating the degree of B-splines (3 for cubic splines)}

\item{boundary_knots}{Vector of two elements specifying the boundary knots}

\item{internal_knots}{Vector containing the desired internal knots of the B-splines}

\item{Results}{List produced from running \code{Sampler_Competition}}

\item{burnin_prop}{Double containing proportion of MCMC samples that should be discarded due to MCMC burn-in (Note burnin_prop includes warm-up iterations)}

\item{time_inhomogeneous}{Boolean containing whether or not a time-inhomogeneous model should be used (if false then basis_degree, boundary_knots, and internal_knots can take any value of the correct type)}

\item{n_samples}{Integer containing number of posterior predictive samples to generate}

\item{trial_time}{Double containing length of trial to simulate}
}
\value{
List containing:
\describe{
  \item{\code{posterior_pred_samples_A}}{Posterior predictive samples of spike trains under the A stimulus}
  \item{\code{posterior_pred_samples_B}}{Posterior predictive samples of spike trains under the B stimulus}
  \item{\code{posterior_pred_samples_AB}}{Posterior predictive samples of spike trains under the A  and B stimuli}
  \item{\code{posterior_pred_labels}}{labels corresponding to posterior_pred_samples_AB}
  \item{\code{n_A}}{Number of spikes in each of the spike trains in posterior_pred_samples_A}
  \item{\code{n_B}}{Number of spikes in each of the spike trains in posterior_pred_samples_B}
  \item{\code{n_AB}}{Number of spikes in each of the spike trains in posterior_pred_samples_AB}
  \item{\code{switch_times}}{Times spent in each encoding state in the spike trains generated in posterior_pred_samples_AB}
  \item{\code{switch_states}}{Encoding state (labels) corresponding to switch_times}
  \item{\code{n_switches}}{number of switches observed in each spike train generated in posterior_pred_samples_AB}
}
}
\description{
This function generates posterior predictive samples of scientific interest from
the competition model. Specifically, this function allows you to obtain (1) posterior
predictive samples from the A condition, B condition, and AB conditions (2) posterior
predictive samples of spike counts under the A condition, B condition, and AB conditions, 
(3) posterior predictive samples of time spent encoding each state (4) posterior predictive 
samples for the number of switches in a trial. This function is to be used after running
\code{Sampler_Competition}.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{burnin_prop}}{must be greater than or equal to 0 and less than 1}
  \item{\code{n_sample}}{must be greater than 1}
}
}

\examples{
##############################
### Time-Homogeneous Model ###
##############################

## Load sample data
dat <- readRDS(system.file("test-data", "time_homogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100

basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain
results <- Sampler_Competition(dat$X_A, dat$X_B, dat$X_AB, dat$n_A, dat$n_B, dat$n_AB, 
                               MCMC_iters, basis_degree, boundary_knots, internal_knots,
                               Warm_block1 = Warm_block1, Warm_block2 = Warm_block2,
                               time_inhomogeneous = FALSE)
                               
## Posterior Predictive Samples
post_pred <- Competition_Posterior_Predictive(1, basis_degree, boundary_knots, internal_knots,
                                              results, time_inhomogeneous = FALSE)

################################
### Time-Inhomogeneous Model ###
################################

## Load sample data
dat <- readRDS(system.file("test-data", "time_inhomogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100
basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain
results <- Sampler_Competition(dat$X_A, dat$X_B, dat$X_AB, dat$n_A, dat$n_B, dat$n_AB, 
                               MCMC_iters, basis_degree, boundary_knots, internal_knots,
                               Warm_block1 = Warm_block1, Warm_block2 = Warm_block2)
                               
## Posterior Predictive Samples                               
post_pred <- Competition_Posterior_Predictive(1, basis_degree, boundary_knots, internal_knots,
                                              results)
                                              
}
