% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FR_CI_IIGPP}
\alias{FR_CI_IIGPP}
\title{Constructs CI for IIGPP Firing Rate}
\usage{
FR_CI_IIGPP(
  time,
  basis_degree,
  boundary_knots,
  internal_knots,
  Results,
  burnin_prop = 0.3,
  alpha = 0.05
)
}
\arguments{
\item{time}{Vector of time points at which pointwise credible intervals will be constructed}

\item{basis_degree}{Integer indicating the degree of B-splines (3 for cubic splines)}

\item{boundary_knots}{Vector of two elements specifying the boundary knots}

\item{internal_knots}{Vector containing the desired internal knots of the B-splines}

\item{Results}{List produced from running \code{Sampler_IIGPP}}

\item{burnin_prop}{Double containing proportion of MCMC samples that should be discarded due to MCMC burn-in}

\item{alpha}{Double indicating the size of the credible interval ((1 - alpha) * 100 percent)}
}
\value{
List containing:
\describe{
  \item{\code{FR_MCMC_Samps}}{Matrix of MCMC samples of the input current function at the specified time points}
  \item{\code{FR_CI}}{Matrix containing the upper and lower values of the credible interval at the specified time points (first column is lower value, second column is upper value)}
  \item{\code{FR_Median}}{Vector containing the estimated posterior median at the specified time points}
}
}
\description{
Constructs credible intervals for the time-inhomgeneous input current function (I) of the inhomogeneous
inverse Gaussian point process (IIGPP) model.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{burnin_prop}}{must be greater than or equal to 0 and less than 1}
  \item{\code{alpha}}{must be between 0 and 0.5}
}
}

\examples{
################################
### Time-Inhomogeneous Model ###
################################

## Load sample data
dat <- readRDS(system.file("test-data", "time_inhomogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100
basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)
end_time <- 1

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain
results <- Sampler_IIGPP(dat$X_A, dat$n_A, MCMC_iters, basis_degree, boundary_knots, 
                         internal_knots, end_time, Warm_block1 = Warm_block1,
                         Warm_block2 = Warm_block2)

## Get CI
time <- seq(0, 1, 0.01)
CI <- FR_CI_IIGPP(time, basis_degree, boundary_knots, internal_knots, results)

## plot CI
CI_range <- c(min(CI$FR_CI)-5, max(CI$FR_CI) + 5)
plot(time, CI$FR_median, type = 'l', col = "red", ylim = CI_range)
lines(time, CI$FR_CI[,1], col = "blue")
lines(time, CI$FR_CI[,2], col = "blue")

}
