% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetBSpline}
\alias{GetBSpline}
\title{Construct B-splines}
\usage{
GetBSpline(time, basis_degree, boundary_knots, internal_knots)
}
\arguments{
\item{time}{Vector of time points of interest}

\item{basis_degree}{Integer indicating the degree of B-splines}

\item{boundary_knots}{Vector of two elements specifying the boundary knots}

\item{internal_knots}{Vector containing the desired internal knots of the B-splines}
}
\value{
Matrix containing the B-splines evaluated at the time points of interest
}
\description{
Creates B-spline basis functions evaluated at the time points of interest.
}
\examples{
time <- seq(0, 1, 0.01)
basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

B <- GetBSpline(time, basis_degree, boundary_knots, internal_knots)

}
