% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMC}
\alias{HMC}
\title{HMC sampler for competition model}
\usage{
HMC(
  Labels,
  X_A,
  X_B,
  X_AB,
  n_A,
  n_B,
  n_AB,
  MCMC_iters,
  Warm_block = 500L,
  init_position = NULL,
  Leapfrog_steps = 10L,
  I_A_shape = 40,
  I_A_rate = 1,
  I_B_shape = 40,
  I_B_rate = 1,
  sigma_A_mean = 6.32,
  sigma_A_shape = 1,
  sigma_B_mean = 6.32,
  sigma_B_shape = 1,
  delta_shape = 0.5,
  delta_rate = 0.1,
  eps_step = NULL,
  step_size = 0.001,
  step_size_delta = 5e-05,
  Mass_mat = NULL
)
}
\arguments{
\item{Labels}{List of vectors containing labels containing membership of spike for AB trials}

\item{X_A}{List of vectors containing ISIs for A trials}

\item{X_B}{List of vectors containing ISIs for B trials}

\item{X_AB}{List of vectors containing ISIs for AB trials}

\item{n_A}{Vector containing number of spikes per trial for A stimulus trials}

\item{n_B}{Vector containing number of spikes per trial for B stimulus trials}

\item{n_AB}{Vector containing number of spikes per trial for AB stimulus trials}

\item{MCMC_iters}{Integer containing number of HMC iterations}

\item{init_position}{Vector containing starting position for HMC}

\item{Leapfrog_steps}{Integer containing number of leapfrog steps for approximating the Hamiltonian dynamics}

\item{I_A_shape}{Double containing shape parameter for prior distribution on I_A}

\item{I_A_rate}{Double containing rate parameter for prior distribution on I_A}

\item{I_B_shape}{Double containing shape parameter for prior distribution on I_B}

\item{I_B_rate}{Double containing rate parameter for prior distribution on I_B}

\item{sigma_A_mean}{Double containing mean parameter for prior distribution on sigma_A}

\item{sigma_A_shape}{Double containing shape parameter for prior distribution on sigma_A}

\item{sigma_B_mean}{Double containing mean parameter for prior distribution on sigma_B}

\item{sigma_B_shape}{Double containing shape parameter for prior distribution on sigma_B}

\item{delta_shape}{Double containing shape parameter for prior distribution on delta}

\item{delta_rate}{Double containing rate parameter for prior distribution on delta}

\item{eps_step}{Vector containing step size for approximating gradient}

\item{step_size}{Vector containing step size for leapfrog integrator}
}
\description{
HMC sampler for competition model
}
