\name{NeuralComp-package}
\alias{NeuralComp-package}
\alias{NeuralComp}
\docType{package}
\title{
\packageTitle{NeuralComp}
}
\description{
\packageDescription{NeuralComp}
}


\author{
\packageAuthor{NeuralComp}

Maintainer: \packageMaintainer{NeuralComp}
}
\references{
Marco, Nicholas, Jennifer M. Groh, and Surya T. Tokdar. "Modeling Neural Switching via Drift-Diffusion Models." arXiv preprint \doi{arXiv:2410.00781} (2024).
}
