% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Sampler_Competition}
\alias{Sampler_Competition}
\title{Sampler for Drift-Diffusion Competition Model}
\usage{
Sampler_Competition(
  X_A,
  X_B,
  X_AB,
  n_A,
  n_B,
  n_AB,
  MCMC_iters,
  basis_degree,
  boundary_knots,
  internal_knots,
  time_inhomogeneous = TRUE,
  Warm_block1 = 500L,
  Warm_block2 = 2000L,
  Leapfrog_steps = 10L,
  I_A_mean = 40,
  I_A_shape = 1,
  I_B_mean = 40,
  I_B_shape = 1,
  sigma_A_mean = 6.32,
  sigma_A_shape = 1,
  sigma_B_mean = 6.32,
  sigma_B_shape = 1,
  delta_shape = 0.01,
  delta_rate = 0.1,
  step_size_theta = 0.001,
  step_size_FR = 0.001,
  delta_proposal_mean = -2,
  delta_proposal_sd = 0.3,
  alpha_labels = 0.2,
  nu = 5,
  gamma = 2,
  delta_adaption_block = 100L,
  Mass_adaption_block = 500L,
  M_proposal = 10L
)
}
\arguments{
\item{X_A}{List of vectors containing the ISIs of A trials}

\item{X_B}{List of vectors containing the ISIs of B trials}

\item{X_AB}{List of vectors containing the ISIs of AB trials}

\item{n_A}{Vector containing number of spikes for each A trial}

\item{n_B}{Vector containing number of spikes for each B trial}

\item{n_AB}{Vector containing number of spikes for each AB trial}

\item{MCMC_iters}{Integer containing the number of MCMC_iterations excluding warm up blocks}

\item{basis_degree}{Integer indicating the degree of B-splines (3 for cubic splines)}

\item{boundary_knots}{Vector of two elements specifying the boundary knots}

\item{internal_knots}{Vector containing the desired internal knots of the B-splines}

\item{time_inhomogeneous}{Boolean containing whether or not a time-inhomogeneous model should be used (if false then basis_degree, boundary_knots, and internal_knots can take any value of the correct type)}

\item{Warm_block1}{Integer containing number of iterations to adapt the leapfrog step size under identity mass matrices}

\item{Warm_block2}{Integer containing number of iterations to adapt the mass matrices, leapfrog step size, and delta sampling parameters}

\item{Leapfrog_steps}{Integer containing number of leapfrog steps per HMC step}

\item{I_A_mean}{Double containing the value for the mean parameter of IG prior on I_A}

\item{I_A_shape}{Double containing the value for the shape parameter of IG prior on I_A}

\item{I_B_mean}{Double containing the value for the mean parameter of IG prior on I_B}

\item{I_B_shape}{Double containing the value for the shape parameter of IG prior on I_B}

\item{sigma_A_mean}{Double containing the value for the mean parameter of IG prior on sigma_A}

\item{sigma_A_shape}{Double containing the value for the shape parameter of IG prior on sigma_A}

\item{sigma_B_mean}{Double containing the value for the mean parameter of IG prior on sigma_B}

\item{sigma_B_shape}{Double containing the value for the shape parameter of IG prior on sigma_B}

\item{delta_shape}{Double containing the value for the shape parameter of gamma prior on delta}

\item{delta_rate}{Double containing the value for the rate parameter of gamma prior on delta}

\item{step_size_theta}{Double containing initial leapfrog step size for theta parameters}

\item{step_size_FR}{Double containing initial leapfrog step size for time-inhomogeneous firing rate parameters}

\item{delta_proposal_mean}{Double containing the value for mean parameter of the lognormal proposal distribution of delta parameter}

\item{delta_proposal_sd}{Double containing the value for sd parameter of the lognormal proposal distribution of delta parameter}

\item{alpha_labels}{Double containing probability that proposed new deltas come from the prior on delta instead of adapted lognormal distribution}

\item{M_proposal}{Integer containing the number of deltas proposed when sampling delta}

\item{alpha}{Double containing the value for the shape parameter of the inverse gamma prior on I_A_sigma squared and I_B_sigma squared}

\item{beta}{Double containing the value for the scale parameter of the inverse gamma prior on I_A_sigma squared and I_B_sigma squared}

\item{delta_adaptation_block}{Integer containing how often the delta sampling parameters should be updated in Warm_block2}

\item{Mass_adaptation_block}{Integer containing how often the Mass Matrix should be updated in Warm_block2}
}
\value{
List containing:
\describe{
  \item{\code{theta}}{Matrix of samples of the theta parameters (I_A, I_B, sigma_A, sigma_B, delta) where each row is an MCMC sample}
  \item{\code{labels}}{List of matrices where each item in the list contains MCMC samples from an AB trial}
  \item{\code{basis_coef_A}}{Matrix containing MCMC samples of the coefficients of the B-splines for the A process (if time-inhomogeneous)}
  \item{\code{basis_coef_B}}{Matrix containing MCMC samples of the coefficients of the B-splines for the B process (if time-inhomogeneous)}
  \item{\code{I_A_sigma_sq}}{Vector of MCMC samples of I_A_sigma squared}
  \item{\code{I_B_sigma_sq}}{Vector of MCMC samples of I_B_sigma squared}
  \item{\code{LogLik}}{Log-likelihood for each MCMC iteration}
  \item{\code{LogPosterior}}{Log-Posterior for each MCMC iteration}
  
}
}
\description{
Conducts MCMC to get posterior samples from the drift-diffusion competition model.
This function can fit a time-homogeneous model, as well as a time-inhomogeneous 
model.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{I_A_mean}}{must be greater than 0}
  \item{\code{I_A_shape}}{must be greater than 0}
  \item{\code{I_B_mean}}{must be greater than 0}
  \item{\code{I_B_shape}}{must be greater than 0}
  \item{\code{sigma_A_mean}}{must be greater than 0}
  \item{\code{sigma_A_shape}}{must be greater than 0}
  \item{\code{sigma_B_mean}}{must be greater than 0}
  \item{\code{sigma_B_shape}}{must be greater than 0}
  \item{\code{delta_shape}}{must be greater than 0}
  \item{\code{delta_rate}}{must be greater than 0}
  \item{\code{step_size_theta}}{must be greater than 0}
  \item{\code{step_size_FR}}{must be greater than 0}
  \item{\code{delta_proposal_sd}}{must be greater than 0}
  \item{\code{alpha_labels}}{must be between 0 and 1}
  \item{\code{alpha}}{must be greater than 0}
  \item{\code{beta}}{must be greater than 0}
}
}

\examples{
##############################
### Time-Homogeneous Model ###
##############################

## Load sample data
dat <- readRDS(system.file("test-data", "time_homogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100

basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain
results <- Sampler_Competition(dat$X_A, dat$X_B, dat$X_AB, dat$n_A, dat$n_B, dat$n_AB, 
                               MCMC_iters, basis_degree, boundary_knots, internal_knots,
                               Warm_block1 = Warm_block1, Warm_block2 = Warm_block2,
                               time_inhomogeneous = FALSE)


################################
### Time-Inhomogeneous Model ###
################################

## Load sample data
dat <- readRDS(system.file("test-data", "time_inhomogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100
basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain
results <- Sampler_Competition(dat$X_A, dat$X_B, dat$X_AB, dat$n_A, dat$n_B, dat$n_AB, 
                               MCMC_iters, basis_degree, boundary_knots, internal_knots,
                               Warm_block1 = Warm_block1, Warm_block2 = Warm_block2)

}
