% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WAIC_IGP}
\alias{WAIC_IGP}
\title{Calculates WAIC for the Inverse Gaussian Renewal Process}
\usage{
WAIC_IGP(
  X_A,
  X_B,
  X_AB,
  n_A,
  n_B,
  n_AB,
  Results_A,
  Results_B,
  Results_AB,
  basis_degree,
  boundary_knots,
  internal_knots,
  time_inhomogeneous = TRUE,
  burnin_prop = 0.5
)
}
\arguments{
\item{X_A}{List of vectors containing the ISIs of A trials}

\item{X_B}{List of vectors containing the ISIs of B trials}

\item{X_AB}{List of vectors containing the ISIs of AB trials}

\item{n_A}{Vector containing number of spikes for each A trial}

\item{n_B}{Vector containing number of spikes for each B trial}

\item{n_AB}{Vector containing number of spikes for each AB trial}

\item{Results_A}{List produced from running \code{Sampler_Competition} for A trials}

\item{Results_B}{List produced from running \code{Sampler_Competition} for B trials}

\item{Results_AB}{List produced from running \code{Sampler_Competition} for AB trials}

\item{basis_degree}{Integer indicating the degree of B-splines (3 for cubic splines)}

\item{boundary_knots}{Vector of two elements specifying the boundary knots}

\item{internal_knots}{Vector containing the desired internal knots of the B-splines}

\item{time_inhomogeneous}{Boolean containing whether or not a time-inhomogeneous model should be used (if false then basis_degree, boundary_knots, and internal_knots can take any value of the correct type)}

\item{burnin_prop}{Double containing proportion of MCMC samples that should be discarded due to MCMC burn-in (Note burnin_prop includes warm-up iterations)}
}
\value{
waic Double containing the value of the Watanabe-Akaike information criterion on the deviance scale
}
\description{
This function calculates the Watanabe-Akaike information criterion (WAIC) for 
the inverse Gaussian renewal process. This function will use the output from
\code{Sampler_IGP} fit for the A, B, and AB data. The WAIC is defined on the 
deviance scale as waic = -2(lppd - p), where lppd is the log pointwise 
predictive density, and p is the effective number of parameters.
}
\section{Warning}{

The following must be true:
\describe{
  \item{\code{basis_degree}}{must be an integer larger than or equal to 1}
  \item{\code{internal_knots}}{must lie in the range of \code{boundary_knots}}
  \item{\code{burnin_prop}}{must be greater than or equal to 0 and less than 1}
}
}

\examples{
##############################
### Time-Homogeneous Model ###
##############################

## Load sample data
dat <- readRDS(system.file("test-data", "time_homogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100

basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain for A trials
results_A <- Sampler_IGP(dat$X_A, dat$n_A, MCMC_iters, basis_degree, boundary_knots,
                         internal_knots, Warm_block1 = Warm_block1, Warm_block2 = Warm_block2,
                         time_inhomogeneous = FALSE)
                       
## Run MCMC chain for B trials
results_B<- Sampler_IGP(dat$X_B, dat$n_B, MCMC_iters, basis_degree, boundary_knots,
                        internal_knots, Warm_block1 = Warm_block1, Warm_block2 = Warm_block2,
                        time_inhomogeneous = FALSE)

## Run MCMC chain for AB trials
results_AB <- Sampler_IGP(dat$X_AB, dat$n_AB, MCMC_iters, basis_degree, boundary_knots,
                          internal_knots, Warm_block1 = Warm_block1, Warm_block2 = Warm_block2,
                          time_inhomogeneous = FALSE)         
## Calculate WAIC
WAIC <- WAIC_IGP(dat$X_A, dat$X_B, dat$X_AB, dat$n_A, dat$n_B, dat$n_AB, results_A,
                 results_B, results_AB, basis_degree, boundary_knots, internal_knots,
                 time_inhomogeneous = FALSE)


################################
### Time-Inhomogeneous Model ###
################################

## Load sample data
dat <- readRDS(system.file("test-data", "time_inhomogeneous_sample_dat.RDS", package = "NeuralComp"))

## set parameters
MCMC_iters <- 100
basis_degree <- 3
boundary_knots <- c(0, 1)
internal_knots <- c(0.25, 0.5, 0.75)

## Warm Blocks should be longer, however for the example, they are short
Warm_block1 = 50
Warm_block2 = 50

## Run MCMC chain for A trials
results_A <- Sampler_IGP(dat$X_A, dat$n_A, MCMC_iters, basis_degree, boundary_knots,
                         internal_knots, Warm_block1 = Warm_block1, Warm_block2 = Warm_block2)
                       
## Run MCMC chain for B trials
results_B<- Sampler_IGP(dat$X_B, dat$n_B, MCMC_iters, basis_degree, boundary_knots,
                        internal_knots, Warm_block1 = Warm_block1, Warm_block2 = Warm_block2)

## Run MCMC chain for AB trials
results_AB <- Sampler_IGP(dat$X_AB, dat$n_AB, MCMC_iters, basis_degree, boundary_knots,
                          internal_knots, Warm_block1 = Warm_block1, Warm_block2 = Warm_block2)

WAIC <- WAIC_IGP(dat$X_A, dat$X_B, dat$X_AB, dat$n_A, dat$n_B, dat$n_AB, results_A,
                 results_B, results_AB, basis_degree, boundary_knots, internal_knots)

}
