\name{NeuralComp-package}
\alias{NeuralComp-package}
\alias{NeuralComp}
\docType{package}
\title{
\packageTitle{NeuralComp}
}
\description{
\packageDescription{NeuralComp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NeuralComp}
\packageIndices{NeuralComp}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NeuralComp}

Maintainer: \packageMaintainer{NeuralComp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
